app/dashboard/requirements.txt
streamlit>=1.50.0
pandas>=2.3.3
plotly>=5.17.0
google-cloud-storage>=3.6.0
google-auth>=2.43.0
pytz>=2023.3
python-dateutil>=2.9.0
pyarrow>=14.0.0
httpx>=0.27.0
PyYAML>=6.0.1
python-dotenv>=1.0.1
